pyproject.toml
[tool.poetry]
name = "tiny"
version = "0.1.0"
description = "A bioinformatics tool for DNA sequence analysis."
authors = ["Bjorn99 <67769176+Bjorn99@users.noreply.github.com>"]
license = "GPL-3.0"
readme = "README.md"
packages = [{include = "tiny"}]

[tool.poetry.dependencies]
python = "^3.12"
typer = "^0.13.1"
rich = "^13.9.4"
pathlib = "^1.0.1"
biopython = "^1.85"
pandas = "^2.2.3"
numpy = "^2.1.3"
pytest = "^8.3.3"
black = "^24.10.0"
isort = "^5.13.2"
flake8 = "^7.1.1"
pysam = "^0.22.1"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"
black = "^24.10"
flake8 = "^7.1.1"
isort = "^5.13.2"

[tool.poetry.scripts]
tiny = "tiny.cli:app"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
